Moving maintenance model to model.dart
